+++ lib/maze/option/processor.rb
@@ -27,8 +27,6 @@ module Maze
 
           # Document server options
           config.document_server_root = options[Maze::Option::DS_ROOT]
-          config.document_server_bind_address = options[Maze::Option::DS_BIND_ADDRESS]
-          config.document_server_port = options[Maze::Option::DS_PORT]
 
           # Logger options
           config.file_log = options[Maze::Option::FILE_LOG]